test(core): cover directory barrel reachability
